+++ src/Trash/ILogJanitor.cs
@@ -1,0 +1,7 @@
+namespace Trash
+{
+    public interface ILogJanitor
+    {
+        void DeleteOldestLogFiles(int numberOfNewestToKeep);
+    }
+}